+++ CHANGELOG.md
@@ -9,6 +9,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 ### Added
 
 * `--fetch-cache=SECONDS`/`-c` option.
+* `--revision`/`-r` option.
 
 ## [v0.2.0] - 2017-03-01